ccd: save AFTER resizing, don't resize on deserialization
Wow, we were saving the full resolution images to disk, and
running the image resizing functions every time on startup,
instead of using the known `600x600` image function.

No wonder it was so slow.

Starting from this commit, (de)serialization for the `Collection`
is stable and (relatively) fast, so now I can actually start
working on other code.