represent audio/reset state as statics (global, infinite lifetime)
Both of these structs represent an object that all entities
should have a read-only view into, but only a certain entity
(Audio, Kernel) should have a write-lock to.

They also are only created once, and live for the entirety
of the program. These structs now being a static with ALL_CAPS
shows the intent more clearly.