gui: fix last_song panic
This needs to be below `if !self.kernel_returned` else we might
use `.walk()` which assumes your keys are valid for the given
`Collection`, which isn't the case when we haven't received
a `Collection` from `Kernel` yet.

Once we do, our state is either reset because it's invalid or
it is kept because it holds valid keys.