kernel: no par_iter(), exit early for CachePath
If the disk is so slow that CachePath can't finish in a
reasonable amount of time, that means the multiple threads will
be hamming away at STDOUT, printing that it is "exiting early",
but since it is a closure, it can't actually exit.

Using `par_iter()` would be nice here, but it means we need to
use a closure, which means we can't _actually_ exit out of the
function early.

So, to prevent thrashing, do this single-threaded
and actually exit if we detect a Collection reset.