resampler: revert https://github.com/hinto-janai/festival/commit/496bf41427e2ac70f4f3d06d19f90c90152be862
The old code was actually correct, we're creating and returning
the interleaved from the `self.output`, so `self.interleaved`
should match the output length, not the other way around.

The reason panics happened before was probably due to the
interleaved having an old length from a previous resample.

Well, Festival GUI v1.3.2 is already out there so ... next time.